// client/src/config/endpoints.ts
export enum Endpoint {
  REGISTER = '/api/auth/register',
  LOGIN = '/api/auth/login',
  CREATE_STICKER_BOOK = '/api/stickerbook/create',
  LOAD_STICKER_BOOKS = '/api/stickerbook',
  CREATE_ARTIST = '/api/artist/create',
  LOAD_ARTISTS = '/api/artist',
  CREATE_POST = '/api/post/create',
  LOAD_POSTS = '/api/post'
}